Modules, Guis: Remove lifetime-dependencies between modules
If Modules/Guis that depend on another Module/Gui store references to
those, their lifetimes have to be well defined. std::shared_ptr did
deliver just that, however I find it much cleaner to pass a non-owning
reference to the calls that actually need it and completely remove the
concerns about lifetime.

The interface of a given Module/Gui does suffer a bit by doing this;
however, the client-code will be a lot clearer since no calls to
std::make_shared() and the like are necessary anymore. Plus, I cannot
really hate on a solution that makes dependencies clearer than before.