Modules, Guis: Remove lifetime-dependency to Config
Explicity passing the Config-structure to the methods that depend on it,
makes the interfaces a bit wonky but reasoning about object-lifetime
becomes a lot simpler.
Until now, the Modules/Guis would store a raw reference which could
dangle. Wrapping every Config-instance in an std::shared_ptr would
solve the problem but introduces dependencies to an implementation
detail (how the lifetime is managed) that does not belong in the
Module/Gui. Passing a non-owning reference to the function calls pushes
the lifetime-concerns into the caller, who can store and move the
objects however they like.

Many Modules/Guis are now stateless and could technically converted into
namespaces. There is, however, a consistency-argument to be made, as
there are a few cases where moving state out of the class becomes
unnatual. Gui/Dialogs and its internal confirmation-timer would be the
best example.