Framework: Move GW::StoC::RegisterPacketCallback out of Instance::ctor
The several RegisterCallback-functions store a std::function objects
within GWCA's internals. Until now, the ctor would pass a lambda that
captures a this-pointer to this function, which can dangle.
By moving these calls out of the class and into a function that only
takes an std::shared_ptr<April::Instance>, object lifetime is now
undebatable.