Framework: Remove Initialization-component
Since the last commit moved most of the initialization logic into the
Instance component, there is not much reason to keep the Initialization
component at all. Its contents are now split up between DllMain and
Instance.

However, this uncovered another upwards dependency from ChatCommands
to Initialization. In order to remove this, the boolean that indicates
whether April should shut down, is now passed into ChatCommands
explicitly.
This may or may not be a step into the right direction but it makes
abundantly clear how all of the passed in modules are basically just
glorified out-parameters. I think this indicates a major design flaw
that runs through the project and needs attention...