Modules, Guis: Move calls into other Modules, Guis, and GW to Framework
Until now, Modules and Guis directly called into one another, which not
only invited circular dependencies but also caused interfaces that feel
unnatural:
A function checking if we want to block a packet should not set a value
in some passed-in GWCA-structure, it should return that decision.
A module reading CLI-like input should not itself call all possible
receivers, but just translate the message into something useful.

This commit removes state-changing calls from one Module/Gui to another
and also calls into GW that are not just reading data from memory.
This lets Modules/Guis act move passively: They now just contain the
specific decision-making logic and leave the actual work (and all the
dependencies that come with it) to a higher level component.

This is implemented by a version of the Command pattern, utilizing
std::variant to avoid inheritance and dynamic allocations. The call
operators are implemented in Instance.
Sadly, there is a massive downside to this: Now all Modules/Guis depend
on Command, which -when changed- will trigger everything to recompile.
A clear dependency-structure should outweigh that, though.

There are two exceptions, where the redesign could not be applied:
    - CursorFix, because Win32-Api demands fixed interfaces
    - Gui/Settings, see below

Technically, every change to a Config in Gui/Settings should constitute
a Command that is passed to Instance and this rewrite will come.
Currently, however, I want to focus on the overall structure of April.